Re-implement models in ActiveRecord on Postgres
Way back when, Mongolab broke and has proven mysteriously hard to reconnect to from Heroku, even though I can take care of it on the command line easily enough. I wanted to get this app working again, and it really should have been on Postgres to begin with, so this takes care of two birds with one stone.

This still needs cleanup (lots of commented-out Mongoid code) and I need to bring back job queueing with Qu/Redis or delayed_job/postgres.